refactor: moved change issue state to new file
