+++ apps/app/components/command-palette/index.ts
@@ -1,2 +1,3 @@
 export * from "./command-pallette";
 export * from "./shortcuts-modal";
+export * from "./change-issue-state";
